backend/requirements.txt
Django
djangorestframework
djongo
psycopg2-binary
gunicorn
requests
sqlparse==0.2.4
django-cors-headers==3.11.0
redis
